pyproject.toml
[project]
name = "it-assistant"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "agents>=1.4.0",
    "fastapi>=0.136.0",
    "flagembedding>=1.3.5",
    "grandalf>=0.8",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.15",
    "langgraph>=1.1.8",
    "loguru>=0.7.3",
    "magic-pdf>=1.3.12",
    "minio>=7.2.20",
    "modelscope>=1.36.0",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pymilvus>=2.6.12",
    "pymilvus-model>=0.3.2",
    "pymongo>=4.17.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "regex>=2026.4.4",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "torchvision>=0.26.0",
    "uvicorn>=0.44.0",
]
